Move api methods to js folder
