+++ src/UI/UserPortal/components/ChatMessage.vue
@@ -83,7 +83,7 @@
 <script lang="ts">
 import type { PropType } from 'vue';
 import { Message, CompletionPrompt } from '@/js/types';
-import api from '~/server/api';
+import api from '@/js/api';
 
 export default {
 	name: 'ChatMessage',
