+++ src/UI/UserPortal/components/ChatSidebar.vue
@@ -94,7 +94,7 @@
 <script lang="ts">
 import type { PropType } from 'vue';
 import { Session } from '@/js/types';
-import api from '~/server/api';
+import api from '@/js/api';
 
 export default {
 	name: 'ChatSidebar',
